// screens/payslips/types.ts
export type TPayslip = {
  id: number
  employee_id: number
  month: string
  year: number
  bank_details_id: number
  payroll_batch_id: number
  date_processed: string
  processed_by: number
  date_certified: string | null
  certified_by: unknown
  date_approved: string
  approved_by: number
  date_paid: string | null
  is_paid: number
  is_sent: number
  payment_frequency: string
  is_frozen: string
  worked_days: number
  worked_hours: number
  absent_days: number
  absent_hours: number
  description: string
  email_sent: number
  address: string | null
  variance_remark: unknown
  temp_house_allowance: number
  lpf: number
  week_day_overtime_hours: number
  non_week_day_overtime_hours: number
  non_week_day_overtime_pay: number
  employer_nhif: number
  consolidated_relief_allowance: number
  nnhf: number
  nigerian_housing_allowance: number
  nigerian_transport_allowance: number
  nsitf: number
  nita: number
  tax_type: string
  conversion_rate: number
  itf: number
  nhif_relief: number
  currency_id: number
  unpaid_leave_days: number
  unpaid_leave_days_amount: number
  unpaid_leave_policy_name: null
  overtime_hrs_topped_up: number
  consolidated_pay: number
  meets_one_third_rule: number
  extra_employee_nssf_amount: number
  posted_to_quick_books: number
  posted_to_wave: number
  posted_to_xero: number
  rest_day_pay: number
  tax_exemption_amount: number
  cumulated_overtime_balance: number
  expected_worked_hours: number
  half_pay_sick_leave_days: number
  half_pay_sick_leave_pay: number
  is_from_recovery: number
  is_from_payroll_rerun: number
  split_nssf: number
  salary_basic: number
  gross: number
  net: number
  paye: number
  nssf: number
  nhif: number
  personal_tax_relief: number
  pension: number
  other_deductions: number
  insurance_relief: number
  voluntary_provident: number
  absent_pay: number
  initial_pay: number
  taxable_income: number
  deduction_before: number
  employer_pension: number
  fringe_tax: number
  low_interest_benefit: number
  mortgage_interest_relief: number
  employer_nssf: number
  helb: number
  week_day_overtime_pay: number
  ewa: number | null
  salary_advance: number | null
  employee_no: string
  id_no: number | null
  pin_no: string | null
  with_nita: boolean
  currency_code: string | null
}

export interface TBatch {
  approved_by: any
  basic: string
  certified_by: any
  company_id: number
  created_at: string
  date_approved: any
  date_certified: any
  date_matched: any
  date_paid: any
  date_posted: any
  date_processed: string
  employer_nhif: string
  employer_nssf: string
  employer_pension: string
  flagged: number
  fringe_tax: string
  from_second_rerun: number
  id: number
  insurance_relief: string
  low_interest_benefit: string
  lpf: string
  matching: number
  month: string
  nigerian_housing_allowance: string
  nigerian_transport_allowance: string
  payment_frequency: string
  period_start: string
  posted_by: any
  process_status: any
  processed_by: number
  recovered: number
  recovered_basic: string
  recovered_gross: string
  recovered_net: string
  rerun_employee_count: any
  split_nssf: number
  tax_exemption_total: string
  total_consolidated_pay: string
  total_extra_employee_nssf_amount: string
  total_half_pay_sick_leave_pay: string
  total_nsitf: string
  total_taxable_income: string
  unpaid_leave_days_amount: string
  updated_at: string
  verified: number
  voluntary_provident: string
  year: number | string
  gross: number
  paye: number
  nhif: number
  nssf: number
  pension: number
  helb: string
  other_deductions: number
  net: number
  skill_levy: string
  cra: string
  nnhf: string
  total_itf: any
  paid: string
  unpaid: string
  exempted: string
  employees_count: string
  meets_one_third_rule_count: string
  not_meet_one_third_rule_count: string
  approved: boolean
  company_name: string
  currency_code: string
}